journal_backend/pyproject.toml
[project]
name = "journal-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Kumar Aman",email = "aman54kumar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "anyio (==4.9.0)",
    "asgiref (==3.8.1)",
    "attrs (==25.3.0)",
    "build (==1.2.2.post1)",
    "cachecontrol (==0.14.2)",
    "cachetools (==5.5.2)",
    "certifi (==2025.1.31)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.1)",
    "cleo (==2.1.0)",
    "colorama (==0.4.6)",
    "crashtest (==0.4.1)",
    "cryptography (==44.0.2)",
    "defusedxml (==0.7.1)",
    "distlib (==0.3.9)",
    "django (==5.2)",
    "django-cors-headers (==4.7.0)",
    "djangorestframework (==3.16.0)",
    "docx2pdf (==0.1.8)",
    "drf-spectacular (==0.28.0)",
    "dulwich (==0.22.8)",
    "fastjsonschema (==2.21.1)",
    "filelock (==3.18.0)",
    "findpython (==0.6.3)",
    "firebase-admin (==6.7.0)",
    "fonttools (==4.57.0)",
    "fpdf2 (==2.8.3)",
    "google-api-core (==2.24.2)",
    "google-api-python-client (==2.167.0)",
    "google-auth (==2.39.0)",
    "google-auth-httplib2 (==0.2.0)",
    "google-cloud-core (==2.4.3)",
    "google-cloud-firestore (==2.20.2)",
    "google-cloud-storage (==3.1.0)",
    "google-crc32c (==1.7.1)",
    "google-resumable-media (==2.7.2)",
    "googleapis-common-protos (==1.70.0)",
    "graphene (==3.4.3)",
    "graphene-django (==3.2.3)",
    "graphql-core (==3.2.6)",
    "graphql-relay (==3.2.0)",
    "grpcio (==1.72.0rc1)",
    "grpcio-status (==1.72.0rc1)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httplib2 (==0.22.0)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "inflection (==0.5.1)",
    "installer (==0.7.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.1.0)",
    "jsonschema (==4.23.0)",
    "jsonschema-specifications (==2025.4.1)",
    "keyring (==25.6.0)",
    "lxml (==5.4.0)",
    "more-itertools (==10.7.0)",
    "msgpack (==1.1.0)",
    "packaging (==25.0)",
    "pbs-installer (==2025.4.9)",
    "pillow (==11.2.1)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.3.7)",
    "poetry (==2.1.2)",
    "poetry-core (==2.1.2)",
    "promise (==2.3)",
    "proto-plus (==1.26.1)",
    "protobuf (==6.31.0rc1)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.2)",
    "pycparser (==2.22)",
    "pyjwt (==2.10.1)",
    "pyparsing (==3.2.3)",
    "pyproject-hooks (==1.2.0)",
    "python-dateutil (==2.9.0.post0)",
    "python-decouple (==3.8)",
    "python-docx (==1.1.2)",
    "python-dotenv (==1.1.0)",
    "pytz (==2025.2)",
    "pywin32 (==310)",
    "pywin32-ctypes (==0.2.3)",
    "pyyaml (==6.0.2)",
    "rapidfuzz (==3.13.0)",
    "referencing (==0.36.2)",
    "requests (==2.32.3)",
    "requests-toolbelt (==1.0.0)",
    "rpds-py (==0.24.0)",
    "rsa (==4.9.1)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlparse (==0.5.3)",
    "text-unidecode (==1.3)",
    "tomlkit (==0.13.2)",
    "tqdm (==4.67.1)",
    "trove-classifiers (==2025.4.28.22)",
    "typing-extensions (==4.13.2)",
    "tzdata (==2025.2)",
    "uritemplate (==4.1.1)",
    "urllib3 (==2.4.0)",
    "virtualenv (==20.30.0)",
    "zstandard (==0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
